refactoring

+++ src/scene/TitleScene.cpp
@@ -15,7 +15,9 @@ TitleScene::~TitleScene() {
 void TitleScene::update() {
     if (CheckHitKey(KEY_INPUT_E)) {
         Parameter parameter;
-        parameter.set(GameScene::ParameterTagLevel, Define::eLevel::Easy);
+        parameter.set(
+            GameScene::ParameterTagLevel,
+            static_cast<int>(Define::eLevel::Easy));
         const bool stackClear = false;
         _implSceneChanged->onSceneChanged(eScene::Game, parameter, stackClear);
         return;
@@ -23,7 +25,9 @@ void TitleScene::update() {
 
     if (CheckHitKey(KEY_INPUT_N)) {
         Parameter parameter;
-        parameter.set(GameScene::ParameterTagLevel, Define::eLevel::Normal);
+        parameter.set(
+            GameScene::ParameterTagLevel,
+            static_cast<int>(Define::eLevel::Normal));
         const bool stackClear = false;
         _implSceneChanged->onSceneChanged(eScene::Game, parameter, stackClear);
         return;
